+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.ui.cart
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.buna.di.viewModel.viewModel
+import com.buna.di.util.viewModel.viewModel
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
