+++ app/src/main/java/woowacourse/shopping/ui/main/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.view.Menu
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.buna.di.viewModel.viewModel
+import com.buna.di.util.viewModel.viewModel
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.ui.cart.CartActivity
